removed confusing sentence in extra credit # 2 (#81)
This sentence confused me a lot, because extra credit # 2 touches a completely different part of the code than extra credit # 1 (or the starting exercise), so you can perfectly build on top of your current code.  
The same exact sentence is written at the end of extra credit # 1, so maybe it was a copy & paste error?